misc: fix #1 and update Gradle and dependencies
This commit updates the Gradle version to 8.10.2 and the Android Gradle Plugin (AGP) to 8.8.0.
It also adds code style configurations to the project and sets the Gradle home directory.
Also Add a preview parameter provider on first parameter inspection tool.
And updates the MultiplicativeConverter dropdown items to use unitName instead of unit fixing the shadowing, as mentioned in #1